requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
SQLAlchemy
python-dotenv
psycopg2-binary
python-dateutil
pytest
pytest-cov
marshmallow
newrelic
gunicorn
configparser
asgiref